Move some icalduration functions to private API
Moved to private API:
* icaldurationtype_as_int
* icaldurationtype_from_int

GH-848
