+++ src/libical-glib/api/i-cal-duration.xml
@@ -90,20 +90,10 @@
         <custom>	g_return_if_fail (duration != NULL &amp;&amp; I_CAL_IS_DURATION (duration));
 	((struct icaldurationtype *)i_cal_object_get_native ((ICalObject *)duration))->seconds = seconds;</custom>
     </method>
-    <method name="i_cal_duration_new_from_int" corresponds="icaldurationtype_from_int" kind="constructor" since="1.0">
-        <parameter type="gint" name="t" comment="The duration in second"/>
-        <returns type="ICalDuration *" annotation="transfer full" comment="The newly created #ICalDuration" />
-        <comment xml:space="preserve">Creates a #ICalDuration from the duration in second.</comment>
-    </method>
     <method name="i_cal_duration_new_from_string" corresponds="icaldurationtype_from_string" kind="constructor" since="1.0">
         <parameter type="const gchar *" name="str" comment="The string representation of the duration"/>
         <returns type="ICalDuration *" annotation="transfer full" comment="The newly created #ICalDuration" />
         <comment xml:space="preserve">Creates a #ICalDuration from the duration in string.</comment>
-    </method>
-    <method name="i_cal_duration_as_int" corresponds="icaldurationtype_as_int" kind="others" since="1.0">
-        <parameter type="ICalDuration *" name="duration" comment="The #ICalDuration to be converted"/>
-        <returns type="gint" comment="The duration in second" />
-        <comment xml:space="preserve">Converts the #ICalDuration to the representation in second.</comment>
     </method>
     <method name="i_cal_duration_as_ical_string" corresponds="icaldurationtype_as_ical_string_r" kind="others" since="1.0">
         <parameter type="ICalDuration *" name="duration" comment="The #ICalDuration to be converted"/>
