+++ src/libical/CMakeLists.txt
@@ -240,6 +240,8 @@ set(
   icaltimezone.c
   icaltimezone.h
   icaltimezoneimpl.h
+  icalduration_p.h
+  icalduration_p.c
   icalduration.h
   icalduration.c
   icalperiod.h
