+++ src/libical/ical_file.cmake
@@ -45,8 +45,7 @@ file(APPEND ${ICAL_FILE_H_FILE} "#endif\n")
 foreach(_current_FILE ${COMBINEDHEADERSICAL})
   file(STRINGS ${_current_FILE} _lines NEWLINE_CONSUME)
   foreach(_currentLINE ${_lines})
-    # do not remove includes for private headers (of the form ical*_p.h)
-    string(REGEX REPLACE "#include \"ical.*[^_p]\\.h\"" "" _currentLINE "${_currentLINE}")
+    string(REGEX REPLACE "#include \"ical.*\\.h\"" "" _currentLINE "${_currentLINE}")
     string(REGEX REPLACE "#include \"config.*\\.h\"" "" _currentLINE "${_currentLINE}")
     if(NOT "${_currentLINE}" STREQUAL "")
       file(APPEND ${ICAL_FILE_H_FILE} "${_currentLINE}\n")
