+++ src/libical/icalduration.h
@@ -41,23 +41,6 @@ struct icaldurationtype {
         0, 0, 0, 0, 0, 0}
 
 /**
- * @brief Creates a new ::icaldurationtype from a duration in seconds.
- * @param t The duration in seconds
- * @return An ::icaldurationtype representing the duration @a t in seconds
- *
- * @par Example
- * ```c
- * // create a new icaldurationtype with a duration of 60 seconds
- * struct icaldurationtype duration;
- * duration = icaldurationtype_from_int(60);
- *
- * // verify that the duration is one minute
- * assert(duration.minutes == 1);
- * ```
- */
-LIBICAL_ICAL_EXPORT struct icaldurationtype icaldurationtype_from_int(int t);
-
-/**
  * @brief Creates a new ::icaldurationtype from a duration given as a string.
  * @param dur The duration as a string
  * @return An ::icaldurationtype representing the duration @a dur
@@ -77,23 +60,6 @@ LIBICAL_ICAL_EXPORT struct icaldurationtype icaldurationtype_from_int(int t);
  * ```
  */
 LIBICAL_ICAL_EXPORT struct icaldurationtype icaldurationtype_from_string(const char *dur);
-
-/**
- * @brief Converts an ::icaldurationtype into the duration in seconds as `int`.
- * @param duration The duration to convert to seconds
- * @return An `int` representing the duration in seconds
- *
- * @par Usage
- * ```c
- * // create icaldurationtype with given duration
- * struct icaldurationtype duration;
- * duration = icaldurationtype_from_int(3532342);
- *
- * // get the duration in seconds and verify it
- * assert(icaldurationtype_as_int(duration) == 3532342);
- * ```
- */
-LIBICAL_ICAL_EXPORT int icaldurationtype_as_int(struct icaldurationtype duration);
 
 /**
  * Converts an icaldurationtype into the iCal format as string.
