+++ src/test/regression.c
@@ -21,6 +21,7 @@
 #include "test-malloc.h"
 #include "libical/ical.h"
 #include "libical/icaldate_p.h"
+#include "libical/icalduration_p.h"
 #include "libicalss/icalss.h"
 #include "libicalvcal/icalvcal.h"
 #include "libicalvcal/vobject.h"
